fix: increase minimum supported Rust version to 1.42.0
The previous MSRV of 1.39.0 was a guess and it is does not support some
of the dependancies of this crate. Increase to 1.42.0 to resolve the
compatibility issues with 1.39.0.

BREAKING CHANGE: increase the MSRV to 1.42.0

The new MSRV is 1.42.0 and the crate will not compile for earlier
version. The only mitigation is to update you Rust version to at least
1.42.0.