fix(lib): generalize the manifest_path accepted by entry functions
The old manifest_path for verify_conditions(), prepare(), and
list_packages() was an Option<PathBuf> which limited the types that
could be passed. Generalize manifest_path to now allow Option<impl
AsRef<Path>> which allows a broader range of types to be passed in.

BREAKING CHANGE: generalize the manifest_path passed to the main library
functions

The manifest_path specified in verify_conditions(), prepare(), and
list_packages() previously supplied enough type inforation to support
type inferance at the call site. The more generalized manifest_path may
cause type inferance to fail in circumstances where it used to suceed.
The fix for this is to add further type information at the call site for
these functions.