feat(lib): implment publish function
Publish the publishable crates from the workspace to crates.io in order
by dependancies. A crate is publishable if its publish field is either
absent (indicating unrestricted publication) or set to just "crates.io".

Change the list_packages function to use the same crate iterating code
as the publish function.