fix(lib): add no-cycles test to verify function
The presence of cycles prevents list_packages() and publish() from
correctly ordering the packages in the workspace based on a
topological sort (the order of the packages within a cycle is not
deterministic). The package graph without any dev_only links is a
DAG so there is at least one proper topological sort ordering of
the packages. A cycle is only possible is at least one of the links
involed in the cycle is a dev_only link.

list_packages() and publish() exclude dev_only links, but due to a
limitation of the guppy crate such links are still taken into account
when constructing the topological sort.

BREAKING CHANGE: verify function treats cycles as an error

Previously the verify function allowed a workspace with a cycle but
such a workspace could not be published due to a limitation of the
way the publish function works. Such workspaces are now rejected earlier
by the verify function. There is no workaround for workspaces that contain
cycles.