ci(release): change git commit strategy
The old strategy for committing version number bumps was to try to push the changed Cargo.toml (and related) files
directly to the branch being released. The semantic-release part of the CI workflow attempted to turn off branch protection
around running semantic-release to allow this to work with branch protection. It didn't work.

The new workflow instead creates a pull request on a new branch to add the changed Cargo.toml (and related) files back
to the branch being release. This will (eventually) be supplemented with a different workflow to automatically approve and
merge this pull request.