fix: add auto merge workflow for bot PR's
The auto merge job will run for various events associated with a PR but
restricted to PR's submitted by "sbosnick-bot". The PR generated at the
end of the release job in the CI workflow will be such a PR.

Note that the secret containing personal access token used by the PR
step in the release job was updated (outside of this commit) as a token
for the sbosnick-bot user.