fix(build): upgrade dependancies (#23)
The older version of guppy had some problems with some version of serde
and some of its other dependencies. This upgrade fixes those problems.

Upgrade MSRV test to 1.45.0 because the latest version of guppy has an (implicit) MSRV of 1.45.0.

Update repository link and install instructions and fix some clippy warnings that had broken the build.