ci: add mergify rule for own pull requests (#31)
Github does not allow someone to approve their own pull request which
has the effect (on this single contributor project) of making the
"approval with success checks" rule imposible to satisfy. The new
"author=sbosnick" rule will allow me to submit pull requests that
mergify handles.

The implication of this, though, is that me submitting a sucessful pull
request to a branch on which semantic release is active will result in a
release of the project (for the types of commit messages that trigger
such releases).