ci: do not treat docs changes as a release trigger
There's an issue here where I don't know how to configure the
release-notes-generator plugin to mirror this setting, so we docs
commits trigger a release with nothing in the release notes.

I looked at the docs again this morning, and it's just not clear to
me how to fix this mismatch. Instead, I'll use `fix(docs)` to trigger
a release for certain docs changes and use the stock angular release
rules.