fix: update next version number in Cargo.lock
The `prepare` semantic-release hook updates the Cargo.toml file with the
next release number. However, the Cargo.lock file is not updated. This
means after each release, when you pull and execute the next `cargo ...`
command, you see a diff in the Cargo.lock file that bumps the version
number to match the version declared in Cargo.toml.

This commit causes the `prepare` semantic-release hook to bump the
crate's next version number in the Cargo.toml and the Cargo.lockfile.

Limitations: this lockfile-bumping feature is not workspace-aware,
so will only bump the lockfile version in the top-level crate.
Actually, this has only been tested in non-workspace crates, but
hopefully this addition is strictly an improvement over the previous
behavior.