fix: use ubuntu-20.04 runner to compile Node.js native module
Closes #167

The `ubuntu-latest` (22) GitHub Actions runner pulls in glibc v2.35[^1],
which is newer than we can expect to exist on many runtime targets
(see #167). It would be helpful if we could link against an older
version of glibc to add compatibility with older runtime targets.

The easiest way to do this seems to be to compile the dynamically-linked
executable on a system with the desired glibc version. Many people use
docker, VMs, or a chroot for this. I tried for a while to find a way to
use this project's Nix flake to specify a reasonably-old glibc version,
but that appears to be beyond my skill level with Nix. I have to time-box
this investigation here. (If you're reading this and you have wisdom to
share, you will find a grateful audience!)

As a workaround, use the oldest Ubuntu GitHub Actions runner, 20.04, to
compile the Node.js native module. This brings in glibc v2.31[^2], which
should be old enough to handle the original request in the issue.

[^1]: https://github.com/actions/runner-images/blob/c26258168863fb6f3f15307411be1665edfac90c/images/linux/Ubuntu2204-Readme.md
[^2]: https://github.com/actions/runner-images/blob/c26258168863fb6f3f15307411be1665edfac90c/images/linux/Ubuntu2004-Readme.md