ci: avoid chicken-and-egg situation in CI
As a part of #500, we are using an older compilation platform to
leverage an older glibc version, to widen the glibc compatibility
range of semantic-release-cargo. This means the first run of CI
on this platform fails to execute a recently-published version of
semantic-release-cargo, since that version was compiled on a newer
version of glibc.

Since semantic-release-cargo is only being used for its `prepare`
function, and semantic-release-cargo is not a Cargo workspace, we can
update the Cargo.{toml,lock} version without semantic-release-cargo to
avoid the chicken-and-egg situation entirely.